Minor: Changed LoadString and mobile warning.
The notification seen when loading Infinie Fun on mobile devices gets in the way and is hard to close, so 've replaced it with an ugly legacy Message instance, which can probably be tapped/clicked through. Also, the LoadString used by the "keepiy" command now references this repository, in case you use that command.