Replaced _G reference with intended variable.
Previously, Xane's MOdel Recreator was using a _G variable, which wasn't guaranteed to be defined. If it wasn't, it would cause a "attempted to do math (add) with nil" error, breaking the script as soon as it would attempt to index instances within a container.

Hopefully, this update fixes that, along with some long-standing typos.