:m: - "noprompt" also hides game favorite prompts.
Infinite Yield might have added support for the newer "favorite this game?" and other prompts since I created this fork, but I'm too lazy to merge the repositories' code and deal with the conflicts right now.

This small commit hides "ThemeProvider", the ScreenGui that contains the favorite request prompt, which some games like to show without any warning! Since its "yes" button is active immediately and the dialog appears so fast, this could cause me (or other exploiters) to accidentally add a game to an alt. account's favorites, which is mostly pointless since these accounts won't have friends and are supposed to hide from other players.

**Cake Off!**, for example, always asks you to favorite it a few minutes after joining a server. It will happily do this, *even if you're decorating a cake* at that exact moment, which feels *unfair*. This addition will slightly improve that technically impressive game.