Made new logo for :balloon: Infinite Fun!
Infinite Yield's logo is classic, but it's one of the leftovers that made Infinite Fun feel indistinct from what it was forked from. Now, it will show a new, cuter logo (with credit to Infinite Yield at the bottom) when its GUI appears.

The logo can also be viewed here; It's used for the readme document.