Remade saveplace for UniversalSynSaveInstance..
Krampus (a pretty powerful PC executor) is sadly dead, at least for now. I replaced Infinite Yield's savegame/saveplace command to use its SavePlace() function instead of "saveinstance()", which functioned differently in that executor.

Since it's probably never returning, I've completely replaced savegame/saveplace with new code, which loads and uses luau/UniversalSynSaveInstance to accurately capture the current place (or "realm" in Royale High terms) to a somewhat large .rbxlx file on your device!

While I could've included the Recreation API as an option here, it's objectively worse than UniversalSynSaveInstance because it can't save unions or terrain. It may be faster, but this could be an issue for for the average exploiter. Also, this requires less code, since I don't have to make sure the API isn't in use or load it if it hasn't been imported using the "loadapi" command in the current session.