Added keyword check to rhsave command.
RH Accessory Preserver is a really useful command/script, but it could be much better, and more user-friendly.

In the current version (v1), it tries to find a target item in your character's EquippedStorage subfolders, and if the player unlocks a dorm room then "tries out" a piece of furniture before tapping the CHECK button, it might start noticing when any new furniture appears in the room (which includes halos).

Although this seemed "intelligent" when I programmed it, I'm starting to notice that it's way too specific, and you have to go out of your way to use it, removing most of your accessories to make it consider letting you save anything.

Since then, I've gotten better, more general ideas for how this script could work. Rather than getting a direct reference to clothes. just let the player pick what they want from a list, then try to find that item using info WHEN they tap the 💾 (save) button.

If I add a new "window" to the GUI, it could expand this script's potential items to complete player characters and their dorms! I just have to...somehow program those menus and completely change the script's logic for when it tries to set up its variables and save the item... :rage2: